Remove explicit configuration of HAProxy `maxconn` at the global and frontend level.
At the global level it is computed based on `ulimit -n` as per https://docs.haproxy.org/3.1/configuration.html#maxconn.

At the frontend level, it defaults to the global `maxconn` if omitted.

Removing `fullconn` from `defaults` removes it from each `backend`. This defaults to 10%
of each the `maxconn` for each `frontend` that routes to the backend. Given each
`server` has much lower `maxconn` it feels fine to remove this.